manage: prevent saving of dates in case there are already registrations for those dates
In case a date would be removed or changed, for which already an existing registration exists
then the seat assignment will fail.

To avoid that we don't allow users to remove a date for which at least one registration
exists. Users can still remove them of no registrations exist, or add new ones.